1.0.0 release

+++ publish_release.sh
@@ -1,9 +1,12 @@
 #!/bin/bash
 
-DLL=ChebsSwordInTheStone/bin/Release/ChebsSwordInTheStone.dll
+RELEASEDIR=ChebsSwordInTheStone/bin/Release
+DLL=$RELEASEDIR/ChebsSwordInTheStone.dll
+LIB=$RELEASEDIR/ChebsValheimLibrary.dll
+BUN=../chebs-necromancy/ChebsNecromancyUnity/Assets/AssetBundles/chebsswordinthestone
 PLUGINS=ChebsSwordInTheStone/Package/plugins
 README=README.md
-#TRANSLATIONS=Translations
+TRANSLATIONS=Translations
 
 VERSION=$1
 
@@ -30,9 +33,23 @@ if [ ! -f "$README" ]; then
     exit 1
 fi
 
+if [ ! -f "$LIB" ]; then
+    echo "Error: $LIB does not exist or is not readable."
+    exit 1
+fi
+
+if [ ! -f "$BUN" ]; then
+    echo "Error: $BUN does not exist or is not readable."
+    exit 1
+fi
+
+
 cp -f "$DLL" "$PLUGINS" || { echo "Error: Failed to copy $DLL"; exit 1; }
 cp -f "$README" "$PLUGINS/../README.md" || { echo "Error: Failed to copy $README"; exit 1; }
-#cp -rf "$TRANSLATIONS" "$PLUGINS/"  || { echo "Error: Failed to copy Translations"; exit 1; }
+cp -rf "$TRANSLATIONS" "$PLUGINS/"  || { echo "Error: Failed to copy Translations"; exit 1; }
+cp -f "$LIB" "$PLUGINS" || { echo "Error: Failed to copy $LIB"; exit 1; }
+cp -f "$BUN" "$PLUGINS" || { echo "Error: Failed to copy $BUN"; exit 1; }
+cp -f "$BUN.manifest" "$PLUGINS" || { echo "Error: Failed to copy $BUN.manifest"; exit 1; }
 
 ZIPDESTINATION="../bin/Release/ChebsSwordInTheStone.$VERSION.zip"
